requirements.txt
bcrypt==3.2.0
cffi==1.15.0
click==8.0.3
colorama==0.4.4
cryptography==35.0.0
ecdsa==0.17.0
Flask==2.0.2
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
pyasn1==0.4.8
pycparser==2.21
python-dotenv==0.19.2
python-jose==3.3.0
rsa==4.7.2
six==1.16.0
SQLAlchemy==1.4.26
Werkzeug==2.0.2
